add: authorize logic in delete task.

+++ app/Http/Controllers/TaskController.php
@@ -2,10 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\DestroyTaskRequest;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Models\Task;
+use Exception;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class TaskController extends Controller
@@ -43,33 +46,49 @@ class TaskController extends Controller
 
     public function update(UpdateTaskRequest $request, Task $task)
     {
-        $validatedData = $request->validated();
-        $task->update([
-            'name' => $validatedData['name'],
-            'description' => $validatedData['description'],
-            'due_date' => $validatedData['dueDate'],
-            'status' => $validatedData['status'],
-            'image_path' => $validatedData['imagePath'],
-            'priority' => $validatedData['priority'],
-            'assigned_user_id' => $validatedData['assignedUserId'],
-            'project_id' => $validatedData['projectId'],
-            'created_by' => Auth::id(),
-            'updated_by' => Auth::id()
-        ]);
-
-        return response()->json([
-            'message' => 'Task updated Successfully!'
-        ], 200);
+        try {
+            $validatedData = $request->validated();
+            $task->update([
+                'name' => $validatedData['name'],
+                'description' => $validatedData['description'],
+                'due_date' => $validatedData['dueDate'],
+                'status' => $validatedData['status'],
+                'image_path' => $validatedData['imagePath'],
+                'priority' => $validatedData['priority'],
+                'assigned_user_id' => $validatedData['assignedUserId'],
+                'project_id' => $validatedData['projectId'],
+                'created_by' => Auth::id(),
+                'updated_by' => Auth::id()
+            ]);
+    
+            return response()->json([
+                'message' => 'Task updated Successfully!'
+            ], 200);
+        } catch (\Exception $e) {
+            return response()->json([
+                'message' => $e->getMessage()
+            ], $e->getCode());
+        }
     }
 
-    public function destroy(Task $task)
+    public function destroy(Request $request, Task $task)
     {
-        $res = $task->delete();
-
-        if($res) {
+        try {
+            if($request->user()->id === $task['assigned_user_id'])
+            {
+                $res = $task->delete();
+    
+                if($res) {
+                    return response()->json([
+                        'message' => 'Task deleted successfully'
+                    ], 200);
+                }
+            }
+            throw new Exception('Unauthenticated.', 401);
+        } catch (\Exception $e) {
             return response()->json([
-                'message' => 'Task deleted Successfully'
-            ], 200);
+                'message' => $e->getMessage()
+            ], $e->getCode());
         }
     }
 }
